convert the SearchComponent to Kotlin
